feat: remove base task list section, move i-base Task list section
